chore: adapted rust-based types for new arch
